SPAR-48 added logic for active language buttons, removed unnecessary comments, improved logic-data download from database
